conver_bot/user_api/requirements.txt
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
mariadb==1.1.11
mysql-connector-python==9.1.0
openai==1.59.7
packaging==24.2
pillow==11.1.0
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
svgwrite==1.4.3
tqdm==4.67.1
Tree==0.2.4
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
